+++ src/main/java/de/blau/android/easyedit/ClosedWaySplittingActionModeCallback.java
@@ -1,15 +1,10 @@
 package de.blau.android.easyedit;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import android.util.Log;
-import android.view.Menu;
 import androidx.annotation.NonNull;
-import androidx.appcompat.view.ActionMode;
-import de.blau.android.R;
 import de.blau.android.exception.OsmIllegalOperationException;
 import de.blau.android.exception.StorageException;
 import de.blau.android.osm.Node;
@@ -22,12 +17,11 @@ import de.blau.android.osm.Way;
  * @author simon
  *
  */
-public class ClosedWaySplittingActionModeCallback extends NonSimpleActionModeCallback {
-    private static final String   DEBUG_TAG      = "ClosedWaySplit...";
-    private final Way             way;
-    private final Node            node;
-    private final Set<OsmElement> nodes          = new HashSet<>();    // nodes that we can use for splitting
-    private boolean               createPolygons = false;
+public class ClosedWaySplittingActionModeCallback extends AbstractClosedWaySplittingActionModeCallback {
+    private static final String DEBUG_TAG      = "ClosedWaySplit...";
+    private final Way           way;
+    private final Node          node;
+    private boolean             createPolygons = false;
 
     /**
      * Construct a new callback for splitting a closed way/polygon
@@ -58,16 +52,6 @@ public class ClosedWaySplittingActionModeCallback extends NonSimpleActionModeCal
     }
 
     @Override
-    public boolean onCreateActionMode(ActionMode mode, Menu menu) {
-        helpTopic = R.string.help_closedwaysplitting;
-        super.onCreateActionMode(mode, menu);
-        mode.setSubtitle(R.string.actionmode_closed_way_split_2);
-        logic.setClickableElements(nodes);
-        logic.setReturnRelations(false);
-        return true;
-    }
-
-    @Override
     public boolean handleElementClick(OsmElement element) { // NOSONAR
         // due to clickableElements, only valid nodes can be clicked
         super.handleElementClick(element);
@@ -91,12 +75,5 @@ public class ClosedWaySplittingActionModeCallback extends NonSimpleActionModeCal
         manager.finish();
         Log.d(DEBUG_TAG, "split failed at element " + (element != null ? element : "null"));
         return true;
-    }
-
-    @Override
-    public void onDestroyActionMode(ActionMode mode) {
-        logic.setClickableElements(null);
-        logic.setReturnRelations(true);
-        super.onDestroyActionMode(mode);
     }
 }
